pyproject.toml
[tool.poetry]
name = "cross-snapshot-attention-gnn"
version = "0.1.0"
description = ""
authors = ["Til Schniese <til.schniese@gmail.com>"]

[tool.poetry.dependencies]
python = "3.10"
numpy = ">=1.15.0,<1.22.2"
networkx = "^2.8.8"
matplotlib = "^3.6.2"
pylint = "^2.15.8"
tqdm = "^4.64.1"
scikit-learn = "^1.2.1"
scipy = ">=1.7.3,<1.8.0"
seaborn = "^0.12.2"
pandas = "^1.5.3"
xgboost = "^1.7.3"
torchvision = "^0.14.1"
dill = "^0.3.6"
pymc3 = "^3.11.5"
arviz = "0.12"
ray = "^2.4.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

